track schema version
If Hibernate gets a cache value with an unexpected shape, i.e. missing
or additional columns, it's very messy. The cache region has no way to
know whether the value it retrieves has the right shape. Hence the
schema version: increment it every time an entity changes shape and the
region will avoid handing back cached values with the wrong version. If
all writers are updated to the new schema version, then eventually all
the cache values will be updated.